feat: setting up config
