+++ go.mod
@@ -2,4 +2,10 @@ module pet-dex-backend/v2
 
 go 1.21.4
 
-require github.com/joho/godotenv v1.5.1 // indirect
+require github.com/stretchr/testify v1.8.4
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
