+++ pkg/mail/email_request.go
@@ -1,0 +1,11 @@
+package mail
+
+type EmailSendRequest struct {
+	From    string
+	To      []string
+	Html    string
+	Subject string
+	Cc      []string
+	Bcc     []string
+	ReplyTo []string
+}
