+++ pkg/mail/interface.go
@@ -1,0 +1,5 @@
+package mail
+
+type MailInterface interface {
+	Send(request *EmailSendRequest) error
+}
